Skip recreating launchd.plist in PropertyListModifier

Former-commit-id: 846e1728797a96a7e9237477ff97354bc4fec3cf

+++ darwin/BuildScripts/PropertyListModifier.swift
@@ -236,17 +236,24 @@ func updatePropertyListWithEntries(_ newEntries: [String : AnyHashable], atPath 
 
 /// Updates the property list by removing the provided keys (if present) or deletes the file if there are now no entries.
 func removePropertyListEntries(forKeys keys: [String], atPath path: URL) throws {
+    // Check if the file name is 'launchd.plist' and return if it is
+    if path.lastPathComponent == "launchd.plist" {
+        print("Skipping the processing of launchd.plist.")
+        return
+    }
+
     let (entries, format) = try readPropertyList(atPath: path)
     for key in keys {
         entries.removeObject(forKey: key)
     }
-    
+
     if entries.count > 0 {
         try writePropertyList(atPath: path, entries: entries, format: format)
     } else {
         try FileManager.default.removeItem(at: path)
     }
 }
+
 
 /// The path of the info property list for this target.
 func infoPropertyListPath() throws -> URL {
